feat: support the subagent lane on Windows
Only the lock-free subagent lane is ported; the tmux teammate lane stays unsupported and fails clean.
