+++ internal/subagent/syncjob_test.go
@@ -7,9 +7,10 @@ import (
 	"strconv"
 	"strings"
 	"testing"
-
-	"github.com/ethanhq/cc-fleet/internal/fingerprint"
 )
+
+// Cross-platform sync-job board + processAlive guard tests. The fake-claude
+// exec cases live in syncjob_unix_test.go.
 
 // ----- a sync run is visible on the board, without leaking its answer -----
 
@@ -19,7 +20,7 @@ func TestRegisterAndFinalizeSyncJob(t *testing.T) {
 	t.Setenv("HOME", t.TempDir())
 
 	// Register a sync job. PID is THIS (alive) process; SettingsPath empty so the
-	// board's StatusFor uses a bare kill(0) and sees it running.
+	// board's StatusFor uses a bare liveness probe and sees it running.
 	jobID := registerSyncJob(Request{Vendor: "glm", JSON: true, LeadSessionID: "lead-sync-1"}, "glm-4.6")
 	if jobID == "" {
 		t.Fatal("registerSyncJob returned an empty job id")
@@ -63,125 +64,7 @@ func TestRegisterAndFinalizeSyncJob(t *testing.T) {
 	}
 }
 
-func TestRun_SyncRecordsBoardJobNoAnswerLeak(t *testing.T) {
-	xdg := t.TempDir()
-	t.Setenv("XDG_CONFIG_HOME", xdg)
-	t.Setenv("HOME", t.TempDir())
-	writeMinimalVendors(t, xdg)
-
-	fakeClaude := writeFakeBin(t, "#!/bin/sh\nprintf '%s' '"+smokeSuccessJSON+"'\nexit 0\n")
-	orig := loadFP
-	loadFP = func() (*fingerprint.Fingerprint, error) {
-		return &fingerprint.Fingerprint{BinaryPath: fakeClaude}, nil
-	}
-	t.Cleanup(func() { loadFP = orig })
-
-	// A SYNCHRONOUS run (no Background). The caller still gets the answer inline.
-	res := Run(Request{Vendor: "glm", Prompt: "hi", JSON: true, LeadSessionID: "lead-run-1"})
-	if !res.OK || res.Result != "SUBAGENT_SMOKE_OK=42" {
-		t.Fatalf("sync run should return the answer to the caller: %+v", res)
-	}
-	if res.LeadSessionID != "lead-run-1" {
-		t.Fatalf("sync Result should carry lead_session_id: %+v", res)
-	}
-	// The returned envelope is unchanged — no JobID stamped, so CLI output parity
-	// holds (board bookkeeping is a pure side channel).
-	if res.JobID != "" {
-		t.Fatalf("sync Result must not carry a JobID: %q", res.JobID)
-	}
-
-	// The board now sees the finished sync job as done — WITHOUT the answer text.
-	jobs, err := ListJobs()
-	if err != nil || len(jobs) != 1 {
-		t.Fatalf("ListJobs: err=%v n=%d", err, len(jobs))
-	}
-	if jobs[0].Status != "done" || jobs[0].Vendor != "glm" {
-		t.Fatalf("finished sync job wrong: %+v", jobs[0])
-	}
-	if jobs[0].LeadSessionID != "lead-run-1" {
-		t.Fatalf("board job should retain lead_session_id: %+v", jobs[0])
-	}
-	if jobs[0].Result != "" {
-		t.Fatalf("board job must not expose the answer: %q", jobs[0].Result)
-	}
-	// The answer text never reaches any job file on disk for a sync run.
-	dir := filepath.Join(xdg, "cc-fleet", jobsDirName)
-	entries, _ := os.ReadDir(dir)
-	for _, e := range entries {
-		data, _ := os.ReadFile(filepath.Join(dir, e.Name()))
-		if strings.Contains(string(data), "SUBAGENT_SMOKE_OK=42") {
-			t.Fatalf("sync job file %s leaked the answer to disk", e.Name())
-		}
-	}
-}
-
-func TestRun_SyncAutoDetectsLeadSession(t *testing.T) {
-	xdg := t.TempDir()
-	t.Setenv("XDG_CONFIG_HOME", xdg)
-	t.Setenv("HOME", t.TempDir())
-	writeMinimalVendors(t, xdg)
-
-	fakeClaude := writeFakeBin(t, "#!/bin/sh\nprintf '%s' '"+smokeSuccessJSON+"'\nexit 0\n")
-	origFP := loadFP
-	loadFP = func() (*fingerprint.Fingerprint, error) {
-		return &fingerprint.Fingerprint{BinaryPath: fakeClaude}, nil
-	}
-	t.Cleanup(func() { loadFP = origFP })
-
-	origDetect := detectLeadSession
-	detectLeadSession = func() string { return "auto-lead-session" }
-	t.Cleanup(func() { detectLeadSession = origDetect })
-
-	res := Run(Request{Vendor: "glm", Prompt: "hi", JSON: true})
-	if !res.OK {
-		t.Fatalf("Run failed: %+v", res)
-	}
-	if res.LeadSessionID != "auto-lead-session" {
-		t.Fatalf("sync Result LeadSessionID = %q, want auto-lead-session", res.LeadSessionID)
-	}
-	jobs, err := ListJobs()
-	if err != nil || len(jobs) != 1 {
-		t.Fatalf("ListJobs: err=%v n=%d", err, len(jobs))
-	}
-	if jobs[0].LeadSessionID != "auto-lead-session" {
-		t.Fatalf("board job LeadSessionID = %q, want auto-lead-session", jobs[0].LeadSessionID)
-	}
-}
-
-func TestRun_ExplicitLeadSessionOverridesAutoDetect(t *testing.T) {
-	xdg := t.TempDir()
-	t.Setenv("XDG_CONFIG_HOME", xdg)
-	t.Setenv("HOME", t.TempDir())
-	writeMinimalVendors(t, xdg)
-
-	fakeClaude := writeFakeBin(t, "#!/bin/sh\nprintf '%s' '"+smokeSuccessJSON+"'\nexit 0\n")
-	origFP := loadFP
-	loadFP = func() (*fingerprint.Fingerprint, error) {
-		return &fingerprint.Fingerprint{BinaryPath: fakeClaude}, nil
-	}
-	t.Cleanup(func() { loadFP = origFP })
-
-	called := false
-	origDetect := detectLeadSession
-	detectLeadSession = func() string {
-		called = true
-		return "auto-lead-session"
-	}
-	t.Cleanup(func() { detectLeadSession = origDetect })
-
-	res := Run(Request{Vendor: "glm", Prompt: "hi", JSON: true, LeadSessionID: "explicit-lead"})
-	if !res.OK {
-		t.Fatalf("Run failed: %+v", res)
-	}
-	if called {
-		t.Fatal("detectLeadSession should not run when LeadSessionID is explicit")
-	}
-	if res.LeadSessionID != "explicit-lead" {
-		t.Fatalf("LeadSessionID = %q, want explicit-lead", res.LeadSessionID)
-	}
-}
-
-// ----- processAlive PID-reuse guard via /proc/<pid>/cmdline -----
+// ----- processAlive PID-reuse guard via the reuseGuardArgv seam -----
 
 func TestProcessAlive_CmdlineReuseGuard(t *testing.T) {
 	// cmdlineIsClaudeJob reads argv through the reuseGuardArgv seam (linux /proc
@@ -253,9 +136,9 @@ func TestProcessAlive_LivePidWrongCmdlineReadsDead(t *testing.T) {
 	if processAlive(os.Getpid(), "/no/such/cc-fleet/profile-marker.json") {
 		t.Fatal("a live pid whose cmdline is not our claude subagent must read as dead (reuse guard)")
 	}
-	// Empty SettingsPath (a sync job / legacy meta) degrades to a bare kill(0).
+	// Empty SettingsPath (a sync job / legacy meta) degrades to a bare liveness probe.
 	if !processAlive(os.Getpid(), "") {
-		t.Fatal("empty settingsPath should degrade to a bare kill(0) = alive for a live pid")
+		t.Fatal("empty settingsPath should degrade to a bare liveness probe = alive for a live pid")
 	}
 	// pid <= 0 is always dead.
 	if processAlive(-1, "") {
