+++ internal/teardown/teardown_recovery_test.go
@@ -1,3 +1,5 @@
+//go:build !windows
+
 package teardown
 
 import (
